+++ src/main/java/step2/READMD.md
@@ -1,4 +1,22 @@
 ### 기능 요구사항
-- 로또 구입 금액을 입력하면 구입 금액에 해당하는 로또를 발급해야 한다.
+- 구매할 Lotto의 매수 구하기
+  - 1000원 -> 1
+  - 1500원 -> 1
+  - 500원  -> error
+  
 - 로또 1장의 가격은 1000원이다.
-- 가격은 변동 될 수 있다.
+  - 가격은 변동 될 수 있다.
+  
+- Lotto 생성
+  - 한 장 생성
+  - 여러 장 생성 
+
+- 당첨 번호 생성
+  - 정상적인 당첨번호 입력
+  - 유효하지 않은 당첨번호
+  
+- 당첨 결과 구하기
+  - 한 장
+  - 여러 장
+  
+- Lotto 결과에 따른 수익률 구하기
