--- src/main/java/step2/io/Input.java
@@ -1,5 +1,0 @@
-package step2.io;
-
-public interface Input {
-    String nextLine();
-}
