--- src/main/java/step2/io/Output.java
@@ -1,5 +1,0 @@
-package step2.io;
-
-public interface Output {
-    void write(String line);
-}
