--- src/main/java/step2/io/StandardInput.java
@@ -1,12 +1,0 @@
-package step2.io;
-
-import java.util.Scanner;
-
-public class StandardInput implements Input {
-    private final Scanner scanner = new Scanner(System.in);
-
-    @Override
-    public String nextLine() {
-        return scanner.nextLine();
-    }
-}
