--- src/main/java/step2/io/StandardOutput.java
@@ -1,8 +1,0 @@
-package step2.io;
-
-public class StandardOutput implements Output {
-    @Override
-    public void write(final String line) {
-        System.out.println(line);
-    }
-}
