--- src/main/java/step2/lotto/LottoConstant.java
@@ -1,9 +1,0 @@
-package step2.lotto;
-
-public final class LottoConstant {
-    public static final int LOTTO_NUMBER_LENGTH = 6;
-    public static final int LOTTO_NUMBER_LOWER_BOUND = 1;
-    public static final int LOTTO_NUMBER_UPPER_BOUND = 45;
-
-    public static final int DEFAULT_LOTTO_PRICE = 1000;
-}
