--- src/main/java/step2/shop/LottoNumberGenerationPolicy.java
@@ -1,7 +1,0 @@
-package step2.shop;
-
-import java.util.List;
-
-public interface LottoNumberGenerationPolicy {
-    List<Integer> nextSix();
-}
