--- src/main/java/step2/statistics/LottoStatistics.java
@@ -1,9 +1,0 @@
-package step2.statistics;
-
-import java.util.List;
-
-import step2.lotto.Lottos;
-
-public interface LottoStatistics {
-
-}
